MSBDS/requirements.txt
aiofiles==22.1.0
aiosqlite==0.19.0
amqp==5.1.1
annotated-types==0.5.0
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.6.0
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
autopep8==1.5.7
Babel==2.12.1
backcall==0.2.0
beautifulsoup4==4.12.2
billiard==4.1.0
bleach==6.0.0
CacheControl==0.14.0
cachetools==5.3.3
celery==5.3.1
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.6
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
comm==0.1.3
contourpy==1.0.7
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.3
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
decouple==0.0.7
defusedxml==0.7.1
deprecation==2.1.0
distlib==0.3.6
Django==4.2.4
django-admin-argon-dashboard==1.0.16
django-admin-datta==1.0.10
django-ajax-selects==2.2.0
django-allauth==0.54.0
django-bootstrap-datepicker-plus==5.0.4
django-bootstrap-icons==0.8.3
django-bootstrap4==23.2
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-chartjs==2.3.0
django-compat==1.0.15
django-cors-headers==3.8.0
django-countries==7.5.1
django-crispy-bootstrap==0.1.1.1
django-crispy-forms==2.0
django-datetimepicker==3.14
django-email-verification==0.3.1
django-money==3.3.0
django-redis==5.3.0
django-tempus-dominus==5.1.2.17
django-timezone-field==5.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
dnspython==2.4.1
entrypoints==0.4
eventlet==0.33.3
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.2
firebase-admin==6.5.0
fonttools==4.39.4
fqdn==1.5.1
geographiclib==2.0
geopy==2.4.1
google==3.0.0
google-api-core==2.19.0
google-api-python-client==2.129.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.1
google-cloud-firestore==2.16.0
google-cloud-storage==2.16.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.63.0
googlemaps==4.10.0
greenlet==2.0.2
grpcio==1.63.0
grpcio-status==1.62.2
gunicorn==20.1.0
httplib2==0.22.0
idna==3.4
ipykernel==6.23.0
ipython==8.13.2
ipython-genutils==0.2.0
ipywidgets==8.0.2
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.6.3
jupyter-ydoc==0.2.4
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.9.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterlab==3.6.3
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
jupyterlab_server==2.22.1
kiwisolver==1.4.4
kombu==5.3.1
lxml==4.9.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
msgpack==1.0.8
mysql-connector-python==8.0.28
mysqlclient==2.2.0
nbclassic==1.0.0
nbclient==0.7.4
nbconvert==7.4.0
nbformat==5.8.0
nest-asyncio==1.5.6
notebook==6.5.4
notebook_shim==0.2.3
numpy==1.24.2
oauthlib==3.2.2
packaging==23.1
palettable==3.3.3
pandas==2.0.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
pipdeptree==2.13.0
platformdirs==3.6.0
prometheus-client==0.16.0
prompt-toolkit==3.0.38
proto-plus==1.23.0
protobuf==4.25.3
psutil==5.9.5
pure-eval==0.2.2
py-moneyed==3.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycodestyle==2.7.0
pycountry==22.3.5
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pypng==0.20220715.0
pyrsistent==0.19.3
python-crontab==3.0.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.3.2
QtPy==2.2.0
redis==4.6.0
requests==2.30.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.0
Send2Trash==1.8.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.4.1
SQLAlchemy==2.0.17
sqlparse==0.4.4
stack-data==0.6.2
terminado==0.17.1
threadpoolctl==3.1.0
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
tornado==6.3.1
traitlets==5.9.0
typing_extensions==4.7.1
tzdata==2023.3
uri-template==1.2.0
uritemplate==4.1.1
urllib3==2.0.2
vine==5.0.0
virtualenv==20.23.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
whitenoise==6.5.0
widgetsnbextension==4.0.3
y-py==0.5.9
ypy-websocket==0.8.2
